API

Molly API

The API reference can be found here. Molly also re-exports StaticArrays.jl, making the likes of SVector available when you call using Molly.

Molly.AndersenThermostat

Molly.Angle

Molly.Angletype

Molly.Atom

Molly.Atomtype

Molly.Bond

Molly.Bondtype

Molly.CoordinateLogger

Molly.Coulomb

Molly.Dihedral

Molly.Dihedraltype

Molly.DistanceNeighbourFinder

Molly.GeneralInteraction

Molly.Interaction

Molly.LennardJones

Molly.Logger

Molly.NeighbourFinder

Molly.NoNeighbourFinder

Molly.NoThermostat

Molly.NoVelocityVerlet

Molly.Simulation

Molly.Simulator

Molly.SpecificInteraction

Molly.StructureWriter

Molly.TemperatureLogger

Molly.Thermostat

Molly.VelocityVerlet

Molly.accelerations

Molly.apply_thermostat!

Molly.find_neighbours

Molly.force!

Molly.log_property!

Molly.maxwellboltzmann

Molly.readinputs

Molly.simulate!

Molly.temperature

Molly.vector

Molly.vector1D

Molly.velocity

Molly.AndersenThermostat — Type.

Rescale random velocities according to the Andersen thermostat.

source

Molly.Angle — Type.

A bond angle between three atoms.

source

Molly.Angletype — Type.

Gromacs angle type.

source

Molly.Atom — Type.

Atom(; <keyword arguments>)

An atom and its associated information. Properties unused in the simulation or in analysis can be left with their default values.

Arguments

attype::AbstractString="": the type of the atom.

name::AbstractString="": the name of the atom.

resnum::Integer=0: the residue number if the atom is part of a polymer.

resname::AbstractString="": the residue name if the atom is part of a polymer.

charge::T=0.0: the charge of the atom, used for electrostatic interactions.

mass::T=0.0: the mass of the atom.

σ::T=0.0: the Lennard-Jones finite distance at which the inter-particle potential is zero.

ϵ::T=0.0: the Lennard-Jones depth of the potential well.

source

Molly.Atomtype — Type.

Gromacs atom type.

source

Molly.Bond — Type.

A bond between two atoms.

source

Molly.Bondtype — Type.

Gromacs bond type.

source

Molly.CoordinateLogger — Type.

Log the coordinates throughout a simulation.

source

Molly.Coulomb — Type.

The Coulomb electrostatic interaction.

source

Molly.Dihedral — Type.

A dihedral torsion angle between four atoms.

source

Molly.Dihedraltype — Type.

Gromacs dihedral type.

source

Molly.DistanceNeighbourFinder — Type.

Find close atoms by distance.

source

Molly.GeneralInteraction — Type.

A general interaction that will apply to all atom pairs.

source

Molly.Interaction — Type.

An interaction between atoms that contributes to forces on the atoms.

source

Molly.LennardJones — Type.

The Lennard-Jones 6-12 interaction.

source

Molly.Logger — Type.

A way to record a property, e.g. the temperature, throughout a simulation.

source

Molly.NeighbourFinder — Type.

A way to find near atoms to save on simulation time.

source

Molly.NoNeighbourFinder — Type.

Placeholder neighbour finder that returns no neighbours.

source

Molly.NoThermostat — Type.

Placeholder thermostat that does nothing.

source

Molly.NoVelocityVerlet — Type.

NoVelocityVerlet()

The velocity-free Verlet integrator, also known as the Störmer method. In this case the velocities given to the Simulator act as the previous step coordinates for the first step.

source

Molly.Simulation — Type.

Simulation(; <keyword arguments>)

The data needed to define and run a molecular simulation. Properties unused in the simulation or in analysis can be left with their default values.

Arguments

simulator::Simulator: the type of simulation to run.

atoms::Vector{<:Any}: the atoms in the simulation.

specific_inter_lists::Dict{String, Vector{<:SpecificInteraction}}=Dict(): the specific interactions in the simulation, i.e. interactions between specific atoms such as bonds or angles.

general_inters::Dict{String, <:GeneralInteraction}=Dict(): the general interactions in the simulation, i.e. interactions between all or most atoms such as electrostatics.

coords::U: the coordinates of the atoms in the simulation. Typically a Vector of SVectors of any dimension and type T, where T is Float64 or Float32.

velocities::U: the velocities of the atoms in the simulation, which should be the same type as the coordinates. The meaning of the velocities depends on the simulator used, e.g. for the NoVelocityVerlet simulator they represent the previous step coordinates for the first step.

temperature::T=0.0: the temperature of the simulation.

box_size::T: the size of the cube in which the simulation takes place.

neighbour_finder::NeighbourFinder=NoNeighbourFinder(): the neighbour finder used to find close atoms and save on computation.

thermostat::Thermostat=NoThermostat(): the thermostat which applies during the simulation.

loggers::Dict{String, <:Logger}=Dict(): the loggers that record properties of interest during the simulation.

timestep::T: the timestep of the simulation.

n_steps::Integer: the number of steps in the simulation.

n_steps_made::Vector{Int}=[]: the number of steps already made during the simulation. This is a Vector to allow the struct to be immutable.

source

Molly.Simulator — Type.

A type of simulation to run, e.g. leap-frog integration or energy minimisation.

source

Molly.SpecificInteraction — Type.

A specific interaction between sets of specific atoms, e.g. a bond angle.

source

Molly.StructureWriter — Type.

Write 3D output structures to the PDB file format throughout a simulation.

source

Molly.TemperatureLogger — Type.

Log the temperature throughout a simulation.

source

Molly.Thermostat — Type.

A way to keep the temperature of a simulation constant.

source

Molly.VelocityVerlet — Type.

VelocityVerlet()

The velocity Verlet integrator.

source

Molly.accelerations — Method.

accelerations(simulation, neighbours; parallel=true)

Calculate accelerations of all atoms using the bonded and non-bonded forces.

source

Molly.apply_thermostat! — Method.

Apply a thermostat to modify a simulation.

source

Molly.find_neighbours — Method.

Update list of close atoms between which non-bonded forces are calculated.

source

Molly.force! — Function.

Update the force for an atom pair in response to a given interation type.

source

Molly.log_property! — Method.

Log a property thoughout a simulation.

source

Molly.maxwellboltzmann — Method.

Draw from the Maxwell-Boltzmann distribution.

source

Molly.readinputs — Method.

Read a Gromacs topology flat file, i.e. all includes collapsed into one file.

source

Molly.simulate! — Method.

simulate!(simulation; parallel=true)
simulate!(simulation, n_steps; parallel=true)
simulate!(simulation, simulator, n_steps; parallel=true)

Run a simulation according to the rules of the given simulator.

source

Molly.temperature — Method.

Calculate the temperature of a system from the kinetic energy of the atoms.

source

Molly.vector — Method.

Displacement between two coordinate values, accounting for the bounding box.

source

Molly.vector1D — Method.

Displacement between two 1D coordinate values, accounting for the bounding box.

source

Molly.velocity — Method.

Generate a random velocity from the Maxwell-Boltzmann distribution.

source