API

Molly API

The API reference can be found here. Molly also re-exports StaticArrays.jl, making the likes of SVector available when you call using Molly.

Molly.AndersenThermostat

Molly.Angle

Molly.Angletype

Molly.Atom

Molly.Atomtype

Molly.Bond

Molly.Bondtype

Molly.CoordinateLogger

Molly.Coulomb

Molly.Dihedral

Molly.Dihedraltype

Molly.DistanceNeighbourFinder

Molly.GeneralInteraction

Molly.Interaction

Molly.LennardJones

Molly.Logger

Molly.NeighbourFinder

Molly.NoNeighbourFinder

Molly.NoThermostat

Molly.Simulation

Molly.Simulator

Molly.SpecificInteraction

Molly.StructureWriter

Molly.TemperatureLogger

Molly.Thermostat

Molly.VelocityVerlet

Molly.apply_thermostat!

Molly.calc_accelerations

Molly.find_neighbours!

Molly.log_property!

Molly.maxwellboltzmann

Molly.readinputs

Molly.simulate!

Molly.temperature

Molly.update_accelerations!

Molly.update_coordinates!

Molly.update_velocities!

Molly.vector

Molly.vector1D

Molly.velocity

Molly.AndersenThermostat — Type.

Rescale random velocities according to the Andersen thermostat.

source

Molly.Angle — Type.

A bond angle between three atoms.

source

Molly.Angletype — Type.

Gromacs angle type.

source

Molly.Atom — Type.

An atom and its associated information.

source

Molly.Atomtype — Type.

Gromacs atom type.

source

Molly.Bond — Type.

A bond between two atoms.

source

Molly.Bondtype — Type.

Gromacs bond type.

source

Molly.CoordinateLogger — Type.

Log the coordinates throughout a simulation.

source

Molly.Coulomb — Type.

The Coulomb electrostatic interaction.

source

Molly.Dihedral — Type.

A dihedral torsion angle between four atoms.

source

Molly.Dihedraltype — Type.

Gromacs dihedral type.

source

Molly.DistanceNeighbourFinder — Type.

Find close atoms by distance.

source

Molly.GeneralInteraction — Type.

A general interaction that will apply to all atom pairs.

source

Molly.Interaction — Type.

An interaction between atoms that contributes to forces on the atoms.

source

Molly.LennardJones — Type.

The Lennard-Jones 6-12 interaction.

source

Molly.Logger — Type.

A way to record a property, e.g. the temperature, throughout a simulation.

source

Molly.NeighbourFinder — Type.

A way to find near atoms to save on simulation time.

source

Molly.NoNeighbourFinder — Type.

Placeholder neighbour finder that does nothing.

source

Molly.NoThermostat — Type.

Placeholder thermostat that does nothing.

source

Molly.Simulation — Type.

The data associated with a molecular simulation.

source

Molly.Simulator — Type.

A type of simulation to run, e.g. leap-frog integration or energy minimisation.

source

Molly.SpecificInteraction — Type.

A specific interaction between sets of specific atoms, e.g. a bond angle.

source

Molly.StructureWriter — Type.

Write 3D output structures to the PDB file format throughout a simulation.

source

Molly.TemperatureLogger — Type.

Log the temperature throughout a simulation.

source

Molly.Thermostat — Type.

A way to keep the temperature of a simulation constant.

source

Molly.VelocityVerlet — Type.

The velocity Verlet integrator.

source

Molly.apply_thermostat! — Method.

Apply a thermostat to modify a simulation.

source

Molly.calc_accelerations — Method.

Calculate accelerations of all atoms using the bonded and non-bonded forces.

source

Molly.find_neighbours! — Method.

Update list of close atoms between which non-bonded forces are calculated.

source

Molly.log_property! — Method.

Log a property thoughout a simulation.

source

Molly.maxwellboltzmann — Method.

Draw from the Maxwell-Boltzmann distribution.

source

Molly.readinputs — Method.

Read a Gromacs topology flat file, i.e. all includes collapsed into one file.

source

Molly.simulate! — Method.

Simulate molecular dynamics.

source

Molly.temperature — Method.

Calculate the temperature of a system from the kinetic energy of the atoms.

source

Molly.update_accelerations! — Function.

Update the accelerations in response to a given interation type.

source

Molly.update_coordinates! — Method.

Update coordinates of all atoms and bound to the bounding box.

source

Molly.update_velocities! — Method.

Update velocities of all atoms using the accelerations.

source

Molly.vector — Method.

Displacement between two coordinate values, accounting for the bounding box.

source

Molly.vector1D — Method.

Displacement between two 1D coordinate values, accounting for the bounding box.

source

Molly.velocity — Method.

Generate a random velocity from the Maxwell-Boltzmann distribution.

source