Molly API

The API reference can be found here.

Molly also re-exports StaticArrays.jl and Unitful.jl, making the likes of SVector and 1.0u"nm" available when you call using Molly.

The visualize function is only available once you have called using GLMakie. Requires.jl is used to lazily load this code when GLMakie.jl is available, which reduces the dependencies of the package.

Molly.AbstractCoupler

Molly.AbstractCutoff

Molly.AndersenThermostat

Molly.Atom

Molly.AtomData

Molly.CellListMapNeighborFinder

Molly.CoordinateLogger

Molly.Coulomb

Molly.CoulombReactionField

Molly.DistanceCutoff

Molly.DistanceNeighborFinder

Molly.DistanceVecNeighborFinder

Molly.GeneralInteraction

Molly.Gravity

Molly.HarmonicAngle

Molly.HarmonicBond

Molly.Interaction

Molly.InteractionList2Atoms

Molly.InteractionList3Atoms

Molly.InteractionList4Atoms

Molly.LennardJones

Molly.Logger

Molly.Mie

Molly.NeighborFinder

Molly.NeighborList

Molly.NeighborListVec

Molly.NoCoupling

Molly.NoCutoff

Molly.NoNeighborFinder

Molly.OpenMMAtomType

Molly.OpenMMForceField

Molly.PeriodicTorsion

Molly.PeriodicTorsionType

Molly.RBTorsion

Molly.ShiftedForceCutoff

Molly.ShiftedPotentialCutoff

Molly.SoftSphere

Molly.SpecificForce2Atoms

Molly.SpecificForce3Atoms

Molly.SpecificForce4Atoms

Molly.SpecificInteraction

Molly.StormerVerlet

Molly.StructureWriter

Molly.System

Molly.TemperatureLogger

Molly.TotalEnergyLogger

Molly.TreeNeighborFinder

Molly.VelocityLogger

Molly.VelocityVerlet

AtomsBase.velocity

Molly.accelerations

Molly.apply_coupling!

Molly.displacements

Molly.distances

Molly.find_neighbors

Molly.find_neighbors

Molly.force

Molly.forces

Molly.kinetic_energy

Molly.log_property!

Molly.mass

Molly.maxwell_boltzmann

Molly.place_atoms

Molly.place_diatomics

Molly.potential_energy

Molly.rdf

Molly.readinputs

Molly.run_loggers!

Molly.setupsystem

Molly.simulate!

Molly.temperature

Molly.total_energy

Molly.ustripvec

Molly.vector

Molly.vector1D

Molly.velocity_autocorr

Molly.visualize

Molly.wrapcoords

Molly.wrapcoordsvec

Molly.AbstractCoupler — Type

A way to keep properties of a system constant. Custom temperature and pressure couplers should sub-type this type.

Molly.AbstractCutoff — Type

A general type of cutoff encoding the approximation used for a potential. Interactions can be parameterized by the cutoff behavior.

Molly.AndersenThermostat — Type

AndersenThermostat(temperature, coupling_const)

Rescale random velocities according to the Andersen thermostat.

Molly.Atom — Type

Atom(; <keyword arguments>)

An atom and its associated information. Properties unused in the simulation or in analysis can be left with their default values. The types used should be bits types if the GPU is going to be used.

Arguments

index::Int: the index of the atom in the system.

charge::C=0.0: the charge of the atom, used for electrostatic interactions.

mass::M=0.0u"u": the mass of the atom.

σ::S=0.0u"nm": the Lennard-Jones finite distance at which the inter-particle potential is zero.

ϵ::E=0.0u"kJ * mol^-1": the Lennard-Jones depth of the potential well.

Molly.AtomData — Type

AtomData(atom_type, atom_name, res_number, res_name)

Data associated with an atom. Storing this separately allows the atom types to be bits types and hence work on the GPU.

Molly.CellListMapNeighborFinder — Type

CellListMapNeighborFinder(; nb_matrix, matrix_14, n_steps, dist_cutoff, x0, unit_cell)

Find close atoms by distance, and store auxiliary arrays for in-place threading. x0 and unit_cell are optional initial coordinates and system unit cell that improve the first approximation of the cell list structure. The unit cell can be provided as a three-component vector of box sides on each direction, in which case the unit cell is considered OrthorhombicCell, or as a unit cell matrix, in which case the cell is considered a general TriclinicCell by the cell list algorithm.

Example

julia> coords
15954-element Vector{SVector{3, Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}}:
 [2.5193063341012127 nm, 3.907448346081021 nm, 4.694954671434135 nm]
 [2.4173958848835233 nm, 3.916034913604175 nm, 4.699661024574953 nm]
 ⋮
 [1.818842280373283 nm, 5.592152965227421 nm, 4.992100424805031 nm]
 [1.7261366568663976 nm, 5.610326185704369 nm, 5.084523386833478 nm]

julia> box_size
3-element SVector{3, Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}} with indices SOneTo(3):
              5.676 nm
             5.6627 nm
             6.2963 nm

julia> neighbor_finder = CellListMapNeighborFinder(
           nb_matrix=s.neighbor_finder.nb_matrix, matrix_14=s.neighbor_finder.matrix_14, 
           n_steps=10, dist_cutoff=1.2u"nm",
           x0 = coords, unit_cell = box_size
       )
CellListMapNeighborFinder{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}, 3, Float64}
  Size of nb_matrix = (15954, 15954)
  n_steps = 10
  dist_cutoff = 1.2 nm

Molly.CoordinateLogger — Type

CoordinateLogger(n_steps; dims=3)

Log the coordinates throughout a simulation.

Molly.Coulomb — Type

Coulomb(; cutoff, nl_only, weight_14, coulomb_const, force_unit, energy_unit)

The Coulomb electrostatic interaction.

Molly.CoulombReactionField — Type

CoulombReactionField(; dist_cutoff, solvent_dielectric, nl_only, weight_14,
                        coulomb_const, force_unit, energy_unit)

The Coulomb electrostatic interaction modified using the reaction field approximation.

Molly.DistanceCutoff — Type

DistanceCutoff(dist_cutoff)

Cutoff that sets the potential and force to be zero past a specified cutoff point.

Molly.DistanceNeighborFinder — Type

DistanceNeighborFinder(; nb_matrix, matrix_14, n_steps, dist_cutoff)

Find close atoms by distance.

Molly.DistanceVecNeighborFinder — Type

DistanceVecNeighborFinder(; nb_matrix, matrix_14, n_steps, dist_cutoff)

Find close atoms by distance in a GPU and differentiable safe manner.

Molly.GeneralInteraction — Type

A general interaction that will apply to all or most atom pairs. Custom general interactions should sub-type this type.

Molly.Gravity — Type

Gravity(; G, nl_only)

The gravitational interaction.

Molly.HarmonicAngle — Type

HarmonicAngle(; th0, cth)

A harmonic bond angle between three atoms.

Molly.HarmonicBond — Type

HarmonicBond(; b0, kb)

A harmonic bond between two atoms.

Molly.Interaction — Type

An interaction between atoms that contributes to forces on the atoms.

Molly.InteractionList2Atoms — Type

InteractionList2Atoms(is, js, inters)
InteractionList2Atoms(inter_type)

A list of specific interactions between two atoms.

Molly.InteractionList3Atoms — Type

InteractionList3Atoms(is, js, ks, inters)
InteractionList3Atoms(inter_type)

A list of specific interactions between three atoms.

Molly.InteractionList4Atoms — Type

InteractionList4Atoms(is, js, ks, ls, inters)
InteractionList4Atoms(inter_type)

A list of specific interactions between four atoms.

Molly.LennardJones — Type

LennardJones(; cutoff, nl_only, lorentz_mixing, weight_14, force_unit, energy_unit, skip_shortcut)

The Lennard-Jones 6-12 interaction. The potential is given by

\[V_{ij}(r_{ij}) = 4\varepsilon_{ij} \left[\left(\frac{\sigma_{ij}}{r_{ij}}\right)^{12} - \left(\frac{\sigma_{ij}}{r_{ij}}\right)^{6}\right]\]

and the force on each atom by

\[\begin{aligned} \vec{F}_i &= 24\varepsilon_{ij} \left(2\frac{\sigma_{ij}^{12}}{r_{ij}^{13}} - \frac{\sigma_{ij}^6}{r_{ij}^{7}}\right) \frac{\vec{r}_{ij}}{r_{ij}} \\ &= \frac{24\varepsilon_{ij}}{r_{ij}^2} \left[2\left(\frac{\sigma_{ij}^{6}}{r_{ij}^{6}}\right)^2 -\left(\frac{\sigma_{ij}}{r_{ij}}\right)^{6}\right] \vec{r}_{ij} \end{aligned}\]

Molly.Logger — Type

A way to record a property, e.g. the temperature, throughout a simulation. Custom loggers should sub-type this type.

Molly.Mie — Type

Mie(; m, n, cutoff, nl_only, lorentz_mixing, force_unit, energy_unit, skip_shortcut)

The Mie generalized interaction. When m equals 6 and n equals 12 this is equivalent to the Lennard-Jones interaction.

Molly.NeighborFinder — Type

A way to find near atoms to save on simulation time. Custom neighbor finders should sub-type this type.

Molly.NeighborList — Type

NeighborList()
NeighborList(n, list)

Structure to contain pre-allocated neighbor lists.

Molly.NeighborListVec — Type

NeighborListVec(n, list)

Structure to contain neighbor lists for broadcasting.

Molly.NoCoupling — Type

NoCoupling()

Placeholder coupler that does nothing.

Molly.NoCutoff — Type

NoCutoff()

Placeholder cutoff that does not alter forces or potentials.

Molly.NoNeighborFinder — Type

NoNeighborFinder()

Placeholder neighbor finder that returns no neighbors. When using this neighbor finder, ensure that nl_only for the interactions is set to false.

Molly.OpenMMAtomType — Type

OpenMMAtomType(class, element, mass, σ, ϵ)

An OpenMM atom type.

Molly.OpenMMForceField — Type

OpenMMForceField(ff_files...)
OpenMMForceField(T, ff_files...)
OpenMMForceField(atom_types, residue_types, bond_types, angle_types,
                    torsion_types, torsion_order, weight_14_coulomb,
                    weight_14_lj)

An OpenMM force field. Read one or more OpenMM force field XML files by passing them to the constructor.

Molly.PeriodicTorsion — Type

PeriodicTorsion(; periodicities, phases, ks)

A periodic torsion angle between four atoms.

Molly.PeriodicTorsionType — Type

PeriodicTorsionType(proper, periodicities, phases, ks)

A periodic torsion type.

Molly.RBTorsion — Type

RBTorsion(; f1, f2, f3, f4)

A Ryckaert-Bellemans torsion angle between four atoms.

Molly.ShiftedForceCutoff — Type

ShiftedForceCutoff(dist_cutoff)

Cutoff that shifts the force to be continuous at a specified cutoff point.

Molly.ShiftedPotentialCutoff — Type

ShiftedPotentialCutoff(dist_cutoff)

Cutoff that shifts the potential to be continuous at a specified cutoff point.

Molly.SoftSphere — Type

SoftSphere(; cutoff, nl_only, lorentz_mixing, force_unit, energy_unit, skip_shortcut)

The soft-sphere potential.

Molly.SpecificForce2Atoms — Type

SpecificForce2Atoms(f1, f2)

Forces on two atoms arising from an interaction.

Molly.SpecificForce3Atoms — Type

SpecificForce3Atoms(f1, f2, f3)

Forces on three atoms arising from an interaction.

Molly.SpecificForce4Atoms — Type

SpecificForce4Atoms(f1, f2, f3, f4)

Forces on four atoms arising from an interaction.

Molly.SpecificInteraction — Type

A specific interaction between sets of specific atoms, e.g. a bond angle. Custom specific interactions should sub-type this type.

Molly.StormerVerlet — Type

StormerVerlet(; <keyword arguments>)

The Störmer-Verlet integrator. In this case the velocities given to the simulator act as the previous step coordinates for the first step. Does not currently work with units or thermostats.

Arguments

dt::T: the time step of the simulation.

coupling::C=NoCoupling(): the coupling which applies during the simulation.

Molly.StructureWriter — Type

StructureWriter(n_steps, filepath, excluded_res=String[])

Write 3D output structures to the PDB file format throughout a simulation.

Molly.System — Type

System(; <keyword arguments>)

A physical system to be simulated. Properties unused in the simulation or in analysis can be left with their default values. atoms, atoms_data, coords and velocities should have the same length. This is a sub-type of AbstractSystem from AtomsBase.jl and implements the interface described there.

Arguments

atoms::A: the atoms, or atom equivalents, in the system. Can be of any type but should be a bits type if the GPU is used.

atoms_data::AD: other data associated with the atoms, allowing the atoms to be bits types and hence work on the GPU.

general_inters::GI=(): the general interactions in the system, i.e. interactions between all or most atoms such as electrostatics. Typically a Tuple.

specific_inter_lists::SI=(): the specific interactions in the system, i.e. interactions between specific atoms such as bonds or angles. Typically a Tuple.

coords::C: the coordinates of the atoms in the system. Typically a vector of SVectors of 2 or 3 dimensions.

velocities::V=zero(coords): the velocities of the atoms in the system.

box_size::B: the size of the box in which the simulation takes place. Typically a SVector of 2 or 3 dimensions.

neighbor_finder::NF=NoNeighborFinder(): the neighbor finder used to find close atoms and save on computation.

loggers::L=Dict{String, Logger}(): the loggers that record properties of interest during a simulation.

force_unit::F=u"kJ * mol^-1 * nm^-1": the unit of force of the system.

energy_unit::E=u"kJ * mol^-1": the unit of energy of the system.

gpu_diff_safe::Bool: whether to use the GPU implementation. Defaults to isa(coords, CuArray).

Molly.TemperatureLogger — Type

TemperatureLogger(n_steps)
TemperatureLogger(T, n_steps)

Log the temperature throughout a simulation.

Molly.TotalEnergyLogger — Type

TotalEnergyLogger(n_steps)

Log the total energy of the system throughout a simulation.

Molly.TreeNeighborFinder — Type

TreeNeighborFinder(; nb_matrix, matrix_14, n_steps, dist_cutoff)

Find close atoms by distance using a tree search.

Molly.VelocityLogger — Type

VelocityLogger(n_steps; dims=3)

Log the velocities throughout a simulation.

Molly.VelocityVerlet — Type

VelocityVerlet(; <keyword arguments>)

The velocity Verlet integrator.

Arguments

dt::T: the time step of the simulation.

coupling::C=NoCoupling(): the coupling which applies during the simulation.

AtomsBase.velocity — Method

velocity(mass, temperature; dims=3)

Generate a random velocity from the Maxwell-Boltzmann distribution.

Molly.accelerations — Function

accelerations(system, neighbors=nothing; parallel=true)

Calculate the accelerations of all atoms using the general and specific interactions and Newton's second law.

Molly.apply_coupling! — Method

apply_coupling!(system, simulator, coupling)

Apply a coupler to modify a simulation. Custom couplers should implement this function.

Molly.displacements — Method

displacements(coords, box_size)

Get the pairwise vector displacements of a set of coordinates, accounting for the periodic boundary conditions.

Molly.distances — Method

distances(coords, box_size)

Get the pairwise distances of a set of coordinates, accounting for the periodic boundary conditions.

Molly.find_neighbors — Function

find_neighbors(system, neighbor_finder, current_neighbors=nothing, step_n=0; parallel=true)

Obtain a list of close atoms in a system. Custom neighbor finders should implement this function.

Molly.find_neighbors — Function

find_neighbors(s::System,
                nf::CellListMapNeighborFinder,
                current_neighbors=nothing,
                step_n::Integer=0;
                parallel::Bool=true)

Find neighbors using CellListMap, without in-place updating. Should be called only the first time the cell lists are built. Modifies the mutable nf structure.

Molly.force — Function

force(inter, coord_i, coord_j, atom_i, atom_j, box_size)

Calculate the force between a pair of atoms due to a given interation type. Custom interaction types should implement this function.

Molly.forces — Function

forces(system, neighbors=nothing; parallel=true)

Calculate the forces on all atoms using the general and specific interactions.

Molly.kinetic_energy — Method

kinetic_energy(s)

Compute the kinetic energy of the system.

Molly.log_property! — Function

log_property!(logger, system, neighbors=nothing, step_n=0)

Log a property of the system thoughout a simulation. Custom loggers should implement this function.

Molly.mass — Method

mass(atom)

The mass of an atom.

Molly.maxwell_boltzmann — Method

maxwell_boltzmann(mass, temperature)

Generate a random speed along one dimension from the Maxwell-Boltzmann distribution.

Molly.place_atoms — Method

place_atoms(n_atoms, box_size, min_dist)

Obtain n_atoms 3D coordinates in a box with sides box_size where no two points are closer than min_dist, accounting for periodic boundary conditions.

Molly.place_diatomics — Method

place_diatomics(n_molecules, box_size, min_dist, bond_length)

Obtain 3D coordinates for n_molecules diatomics in a box with sides box_size where no two points are closer than min_dist and the bond length is bond_length, accounting for periodic boundary conditions.

Molly.potential_energy — Function

potential_energy(s, neighbors=nothing)

Compute the potential energy of the system. If the interactions use neighbor lists, the neighbors should be computed first and passed to the function.

Molly.rdf — Method

rdf(coords, box_size; npoints=200)

Get the radial distribution function of a set of coordinates. This describes how density varies as a function of distance from each atom. Returns a list of distance bin centres and a list of the corresponding densities.

Molly.readinputs — Method

readinputs(topology_file, coordinate_file; units=true)
readinputs(T, topology_file, coordinate_file; units=true)

Read a Gromacs topology flat file, i.e. all includes collapsed into one file, and a Gromacs coordinate file. Returns the atoms, atoms data, specific interaction lists, general interaction lists, neighbor finder, coordinates and box size. units determines whether the returned values have units.

Molly.run_loggers! — Function

run_loggers!(system, neighbors=nothing, step_n=0)

Run the loggers associated with the system.

Molly.setupsystem — Method

setupsystem(coord_file, force_field; dist_cutoff=1.0u"nm")

Read a coordinate file and apply a force field to it. Any file format readable by Chemfiles can be given. Returns the atoms, specific interaction lists, general interaction lists, neighbor finder, coordinates and box size.

Molly.simulate! — Method

simulate!(system, simulator, n_steps; parallel=true)
simulate!(system, simulator; parallel=true)

Run a simulation on a system according to the rules of the given simulator. Custom simulators should implement this function.

Molly.temperature — Method

temperature(system)

Calculate the temperature of a system from the kinetic energy of the atoms.

Molly.total_energy — Function

total_energy(s, neighbors=nothing)

Compute the total energy of the system.

Molly.ustripvec — Method

ustripvec(x)

Broadcasted form of ustrip from Unitful.jl, allowing e.g. ustripvec.(coords).

Molly.vector — Method

vector(c1, c2, box_size)

Displacement between two coordinate values from c1 to c2, accounting for the bounding box. The minimum image convention is used, so the displacement is to the closest version of the coordinates accounting for the periodic boundaries.

Molly.vector1D — Method

vector1D(c1, c2, side_length)

Displacement between two 1D coordinate values from c1 to c2, accounting for the bounding box. The minimum image convention is used, so the displacement is to the closest version of the coordinate accounting for the periodic boundaries.

Molly.velocity_autocorr — Function

velocity_autocorr(vl, first_ind, last_ind)

Calculates the autocorrelation function of velocity from the velocity logger. This helps characterize the similarity between velocities observed at different time instances.

Molly.visualize — Function

visualize(coord_logger, box_size, out_filepath; <keyword arguments>)

Visualize a simulation as an animation. This function is only available when GLMakie is imported. It can take a while to run, depending on the length and size of the simulation.

Arguments

connections=Tuple{Int, Int}[]: pairs of atoms indices to link with bonds.

connection_frames: the frames in which bonds are shown. Is a list of the same length as the number of frames, where each item is a list of Bools of the same length as connections. Defaults to always true.

trails::Integer=0: the number of preceding frames to show as transparent trails.

framerate::Integer=30: the frame rate of the animation.

color=:purple: the color of the atoms. Can be a single color or a list of colors of the same length as the number of atoms.

connection_color=:orange: the color of the bonds. Can be a single color or a list of colors of the same length as connections.

markersize=20.0: the size of the atom markers.

linewidth=2.0: the width of the bond lines.

transparency=true: whether transparency is active on the plot.

kwargs...: other keyword arguments are passed to the plotting function.

Molly.wrapcoords — Method

wrapcoords(c, side_length)

Ensure a 1D coordinate is within the simulation box and return the coordinate.

Molly.wrapcoordsvec — Method

wrapcoordsvec(c, box_size)

Ensure a coordinate is within the simulation box and return the coordinate.